chore: include test to validate node and edge labels against schema
